Code examples for UPC/EAN article
