+++ A002LuhnAlgorithm/CheckDigits.Tests.Benchmarks/GlobalUsings.cs
@@ -1,5 +1,6 @@
 ﻿global using BenchmarkDotNet.Attributes;
 global using BenchmarkDotNet.Running;
 
+global using CheckDigits.EanAlgorithmVersions;
 global using CheckDigits.Tests.Benchmarks;
