+++ A002LuhnAlgorithm/CheckDigits.Tests.Benchmarks/Program.cs
@@ -1,1 +1,1 @@
-﻿BenchmarkRunner.Run<LuhnAlgorithmBenchmarks>();
+﻿BenchmarkRunner.Run<BarcodeAlgorithmBenchmarks>();
