+++ A002LuhnAlgorithm/CheckDigits.Tests.Unit/GlobalUsings.cs
@@ -1,2 +1,6 @@
 global using FluentAssertions;
 global using Xunit;
+
+
+global using CheckDigits.EanAlgorithmVersions;
+global using CheckDigits.LuhnAlgorithmVersions;
